requirements.txt
asgiref==3.8.1
boto3==1.34.127
botocore==1.34.127
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
jmespath==1.0.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.43.0
whitenoise==6.6.0
